Refactored all commands, splitting them into a separated file for each command
