+++ src/command/TicTacToe/leaderboard.ts
@@ -1,0 +1,6 @@
+import { CommandInteraction } from "discord.js";
+
+export async function leaderboard(interaction: CommandInteraction<"cached">) {
+	// TODO implement leaderboard
+	return interaction.reply("WIP");
+}
