+++ src/command/instanciate-command.ts
@@ -1,10 +1,10 @@
 import { Command } from "./Command.js";
-import { GuildCommand } from "../GuildCommand.js";
-import { GuildLog } from "../GuildLog.js";
-import { GuildAutoRole } from "../GuildAutoRole.js";
-import { GuildEmbedCommand } from "../GuildEmbedCommand.js";
-import { Purge } from "../Purge.js";
-import { TicTacToe } from "../TicTacToe.js";
+import { GuildCommand } from "./GuildCommand/GuildCommand.js";
+import { GuildLog } from "./GuildLog/GuildLog.js";
+import { GuildAutoRole } from "./GuildAutoRole/GuildAutoRole.js";
+import { GuildEmbedCommand } from "./GuildEmbedCommand/GuildEmbedCommand.js";
+import { Purge } from "./Purge/Purge.js";
+import { TicTacToe } from "./TicTacToe/TicTacToe.js";
 
 /**
  * CommandManager will read this list to create all commands,
